Create Hollow-rectangle.c
This repository contains a simple C program that prints a hollow rectangle. The program takes two integer inputs, the number of rows and the number of columns, and prints a rectangle of asterisks with spaces in between. For example, if the user inputs 3 rows and 4 columns, the program will print the following:
* * * *
*      *
* * * *

The program uses a nested for loop to iterate over the rows and columns of the rectangle. In each iteration, the program checks if the current row or column is the first or last row or column. If it is, the program prints an asterisk. Otherwise, the program prints a space.

This program is a simple example of how to use C to print a shape. It can be used as a starting point for more complex programs that print other shapes, such as circles, triangles, and squares.